Update pedestrian_crossing_system.s
This is a final project coded entirely in assembly programming language in order to simulate a real-world pedestrian crossing system. The hardware platform used in this case is the raspberry pi 3 (capitalizing on its GPIO pins, 3.5V power supply, the Raspian OS). Other  miscellaneous hardwares used are the resistances, LEDs, and push button came with the Raspberry Pi Kit. If you have a raspberry pi 3 kit and want to try out this code yourself, just know the "pins" here stand for the LEDs (2 colors,  green = GO, red = STOP, essentially simulating the traffic lights). Don't forget to add the necessary resistance to prevent excessive amperage across the LEDs, and a push button as the "launch button" for the system as what you would usually do when cross an actual street. Hope this small project helps you get a little taste on coding in assembly language (if you code this in C++, it should take no more than 20-30 lines of code actually  =)   ).